requirements.txt
torch>=1.11.0
torchvision>=0.11.0

lightning-bolts
matplotlib
numpy
scipy
scikit-learn
scprep
scanpy
torchdyn>=1.0.6   # 1.0.4 is broken on pypi
pot
torchdiffeq
absl-py
clean-fid
argparse

neuraloperator
pygpc
xgboost
torchsummary
kagglehub
wandb
